Get paths to useful files from the database.
I used to have a list of file-paths to every useful file (for spreadsheets or exported data csvs), and I would load that file in and work with the paths as needed. But this workflow requires me to synchronize the list-file if I worked another computer. So I'm now storing the paths in a table in the database.